requirements.txt
flask
PyPDF2
flask_cors
reportlab
supabase
pikepdf
python-dotenv